refactor: cards list to handle array of items
