+++ app/components/footer.tsx
@@ -1,4 +1,5 @@
 import { FunctionComponent } from 'react'
+import { Link } from 'remix'
 
 interface FooterProps {}
 
@@ -6,13 +7,18 @@ export const Footer: FunctionComponent<FooterProps> = () => {
   return (
     <footer className="footer footer-center p-10 bg-base-200 text-base-content rounded">
       <div className="grid grid-flow-col gap-4">
-        <a className="link link-hover">Beranda</a>
-        <a className="link link-hover">Katalog</a>
-        <a className="link link-hover">Artikel</a>
-        <a className="link link-hover">Pemesanan</a>
+        <Link to="/" className="link link-hover">
+          Beranda
+        </Link>
+        <Link to="/produk" className="link link-hover">
+          Produk
+        </Link>
+        <Link to="/pemesanan" className="link link-hover">
+          Pemesanan
+        </Link>
       </div>
       <div>
-        <p>Copyright © 2022 - Hak Cipta dilindungi oleh Fujibox </p>
+        <p>Copyright © 2022 - Hak Cipta dilindungi oleh Fujibox</p>
       </div>
     </footer>
   )
