+++ app/components/index.ts
@@ -3,6 +3,5 @@ export * from './footer'
 export * from './hero'
 export * from './layout'
 export * from './navbar'
-export * from './product-cards'
-export * from './product-category-section'
+export * from './cards-list'
 export * from './whatsapp-card'
