+++ app/routes/index.tsx
@@ -1,14 +1,6 @@
 import { gql } from '@urql/core'
-import { LoaderFunction, useLoaderData } from 'remix'
-import {
-  ArticleList,
-  ProductCategorySection,
-  Footer,
-  Hero,
-  Navbar,
-  ProductCards,
-  WhatsAppCard,
-} from '~/components'
+import { Link, LoaderFunction, useLoaderData } from 'remix'
+import { CardsList, Hero, WhatsAppCard } from '~/components'
 import { graphcmsClient } from '~/lib'
 
 export const loader: LoaderFunction = async () => {
@@ -55,10 +47,37 @@ export default function Index() {
   return (
     <>
       <Hero />
-      <ProductCards products={products} />
+
+      <div className="container lg mb-8">
+        <div className="flex items-center gap-4 mb-4">
+          <h1 className="text-4xl font-bold">Produk terbaru</h1>
+          <Link className="btn btn-primary btn-sm" to="/produk">
+            Lihat semua produk
+          </Link>
+        </div>
+        <CardsList route="produk" items={products} />
+      </div>
+
       <WhatsAppCard />
-      <ProductCategorySection categories={categories} />
-      {/* <ArticleList /> */}
+
+      <div className="container mx-auto my-20">
+        <div className="flex w-full justify-between flex-col gap-4">
+          <div className="flex flex-col">
+            <div className="flex items-center gap-4 mb-4">
+              <h1 className="text-4xl font-bold">Kategori produk</h1>
+              <Link className="btn btn-primary btn-sm" to="/kategori">
+                Lihat semua kategori
+              </Link>
+            </div>
+            <img
+              className="rounded-2xl object-cover max-h-[520px]"
+              alt="Kategori produk"
+              src="https://media.graphassets.com/ifohFFVvQ26HWzskzf8p"
+            />
+          </div>
+          <CardsList route="kategori" items={categories} />
+        </div>
+      </div>
     </>
   )
 }
