+++ app/routes/produk.tsx
@@ -1,6 +1,6 @@
-import { json, Link, LoaderFunction, useLoaderData } from 'remix'
+import { json, LoaderFunction, useLoaderData } from 'remix'
 import { gql } from '@urql/core'
-import { Footer, Navbar } from '~/components'
+import { CardsList } from '~/components'
 import { graphcmsClient } from '~/lib'
 import { TProduct } from '~/types'
 
@@ -36,15 +36,7 @@ export default function Products() {
   return (
     <>
       <div>
-        {products.map((product) => {
-          return (
-            <div key={product.id}>
-              <Link to={`/produk/${product.slug}`}>{product.name}</Link>
-            </div>
-          )
-        })}
-        {/* 
-        <pre>{JSON.stringify(products, null, 2)}</pre> */}
+        <CardsList route="produk" items={products} />
       </div>
     </>
   )
